+++ src/dictionary/DefaultEditDate.ts
@@ -2,23 +2,28 @@ import DefaultEdit, { DefaultEditInitOption } from "./DefaultEdit"
 import DictionaryValue from "./DictionaryValue"
 
 type disabledDateConfig = { start?: unknown, end?: unknown }
+type disabledTimeConfig = { start?: unknown, end?: unknown }
 
 export interface DefaultEditDateOption {
   format: string
+  hideClear: boolean
   time?: {
     format: string
     defaultValue: string
   }
   disabledDate?: (value: unknown) => boolean
+  disabledTime?: (value: unknown) => boolean
 }
 
 export interface PartialDefaultEditDateOption {
   format?: string
+  hideClear?: boolean
   time?: {
     format?: string
     defaultValue?: string
   }
   disabledDate?: disabledDateConfig | ((value: unknown) => boolean)
+  disabledTime?: disabledTimeConfig | ((value: unknown) => boolean)
 }
 
 export interface DefaultEditDateInitOption extends DefaultEditInitOption {
@@ -30,11 +35,17 @@ class DefaultEditDate extends DefaultEdit{
   static $name = 'DefaultEditDate'
   static $defaultOption = {
     format: 'YYYY-MM-DD',
+    hideClear: false,
     time: {
       format: 'HH:mm:ss',
       defaultValue: '00:00:00'
     },
     disabledDate (option: disabledDateConfig) {
+      return function(value: unknown) {
+        return false
+      }
+    },
+    disabledTime (option: disabledDateConfig) {
       return function(value: unknown) {
         return false
       }
@@ -48,7 +59,8 @@ class DefaultEditDate extends DefaultEdit{
     const option = initOption.option || {}
     const $defaultOption = (this.constructor as typeof DefaultEditDate).$defaultOption
     this.$option = {
-      format: option.format || $defaultOption.format
+      format: option.format || $defaultOption.format,
+      hideClear: option.hideClear === undefined ? $defaultOption.hideClear : option.hideClear
     }
     if (option.time) {
       this.$option.time = {
@@ -59,6 +71,11 @@ class DefaultEditDate extends DefaultEdit{
     if (option.disabledDate) {
       if (typeof option.disabledDate === 'object') {
         this.$option.disabledDate = $defaultOption.disabledDate(option.disabledDate)
+      }
+    }
+    if (option.disabledTime) {
+      if (typeof option.disabledTime === 'object') {
+        this.$option.disabledTime = $defaultOption.disabledTime(option.disabledTime)
       }
     }
   }